# docs/searches-api/api.yaml
swagger: "2.0"
info:
  description: "This a API docs for Searches API of RaptureReady.tv"
  version: "1.0.0"
  title: "Searches API"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "raptureready@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "http://www.raptureready.tv"
basePath: "/api"
schemes:
- "http"
paths:
  /channel/{id}/related:
    get:
      summary: "Get related channels"
      description: "Get some related channels of any channel"
      operationId: "getRelatedChannels"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "id"
        name: "id"
        description: "Channel id"
        required: true
      responses:
        200:
          description: "Channels list"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Channel"
  /channel/?q={query}:
    get:
      summary: "Search channels"
      description: "Search channels by title, description or slug"
      operationId: "searchChannelsByQuery"
      produces:
      - "application/json"
      parameters:
      - name: "query"
        in: "query"
        description: "Keyword to search"
        required: true
        type: "string"
      responses:
        200:
          description: "Channels list"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Channel"
  /channel/?filter={letter}:
    get:
      summary: "Filter channels"
      description: "Filter channels by their first letter"
      operationId: "filterChannelsByLetter"
      produces:
      - "application/json"
      parameters:
      - name: "letter"
        in: "letter"
        description: "Letter to filter"
        required: true
        type: "string"
      responses:
        200:
          description: "Channels list"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Channel"
definitions:
  Channel:
    type: "object"
    properties:
      _id:
        type: "string"
      name:
        type: "string"
      slug:
        type: "string"
      channelNumber:
        type: "integer"
      title:
        type: "string"
      description:
        type: "string"
      thumb:
        type: "string"
      rating:
        type: "integer"
      age:
        type: "string"
      type:
        type: "string"
      embedCode:
        type: "string"
      fix:
        type: "boolean"
      comment:
        type: "string"
      disabled:
        type: "boolean"
      views:
        type: "integer"
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"